Refactor
Moved settings areas into their own files
